Working on Add Customer Screen
Created models for country and division.
Added new queries for first-level division and countries
Code is set up to accept user data and add customer to database.
Username is carried over from login screen to use in adding a customer.
Combo boxes for 1st level division and country are functional. Currently there is an error when trying to select a different country to update 1st level division; program crashes.